+++ webapp/src/views/organism/form-builder/thankyou-page.tsx
@@ -9,13 +9,19 @@ import { IsValidString } from '@app/utils/string-utils';
 import GreetingLayoutWrapper from '../layout/greeting-layout-wrapper';
 
 const ThankYouSlide = ({ disabled }: { disabled?: boolean }) => {
-    const { formState, theme, setThankYouPageDescription } = useFormState();
+    const { formState, theme, setThankYouPageTitle, setThankYouPageDescription } = useFormState();
     const { activeThankYouPageComponent } = useActiveThankYouPageComponent();
     return (
         <GreetingLayoutWrapper disabled={disabled} greetingIndex={-20} theme={theme}>
             <div className={cn('flex w-full  items-start', formState.thankyouPage![activeThankYouPageComponent?.index || 0]?.layout === FormSlideLayout.SINGLE_COLUMN_NO_BACKGROUND_LEFT_ALIGN ? 'justify-start' : ' justify-center')}>
                 <div className="flex w-full max-w-[800px] flex-col items-start">
-                    <h1 className="text-[40px] font-semibold">Thank You! 🎉</h1>
+                    <AutosizeTextarea
+                        style={{ resize: 'none' }}
+                        placeholder="Thank You! 🎉"
+                        className="w-full border-0 px-0 py-0 text-[40px] font-semibold leading-[48px]"
+                        value={formState.thankyouPage![activeThankYouPageComponent?.index || 0]?.title ?? ''}
+                        onChange={(event: any) => setThankYouPageTitle(activeThankYouPageComponent?.index || 0, event.target.value)}
+                    />
                     {formState.thankyouPage && IsValidString(formState.thankyouPage[activeThankYouPageComponent?.index || 0].message) ? (
                         <AutosizeTextarea
                             placeholder="Your response has been successfully submitted"